2024/pyproject.toml
[tool.poetry]
name = "advent-2022"
version = "0.1.0"
description = ""
authors = ["Pieter Vanderpol <pjvanderpol@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
setuptools = "^65.6.3"
numpy = "^1.24.0"
ipython = "^8.7.0"
pandas = "^1.5.2"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
pybars3 = "^0.9.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
